requirements-dev.txt
flake8==4.0.0
autoflake==1.4
pre-commit==2.19.0
pytest==7.1.2
pytest-cov==3.0.0
black==22.6.0
isort==5.10.1
mypy==0.961
pandas-profiling==3.2.0